% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jstemplate.R
\name{generateOptionsJSCode}
\alias{generateOptionsJSCode}
\title{Generate javascript code based on general options, options
for containers and network data}
\usage{
generateOptionsJSCode(opts)
}
\arguments{
\item{opts}{\code{list} containing general options for GraphRender
component}
}
\value{
String containing JS code for component options
}
\description{
Generate javascript code based on general options, options
for containers and network data
}
\author{
Sylvain Gubian \email{DL.RSupport@pmi.com}
}

