% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataformating.R
\name{getAdjMat}
\alias{getAdjMat}
\title{Get a RGraph2js compatible adjacency matrix from the provided R object.}
\usage{
getAdjMat(A)
}
\arguments{
\item{A}{signed weighted adjacency \code{matrix} or an instance of the class
\code{graphAM}, \code{graphBAM}, \code{graphNEL} or \code{clusterGraph}
from the graph package}
}
\value{
the RGraph2js compatible adjacency \code{matrix}
}
\description{
Get a RGraph2js compatible adjacency matrix from the provided R object.
}
\author{
Stephane Cano \email{stephane.cano@pmi.com}, PMP SA.
}

