% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmltemplate.R
\name{getDefaultToolParameters}
\alias{getDefaultToolParameters}
\title{Function wich generates a list containing parameters for Tools in
the D3js component with default values}
\usage{
getDefaultToolParameters()
}
\value{
\code{list} of parameters with default values
}
\description{
Function wich generates a list containing parameters for Tools in
the D3js component with default values
}
\author{
Sylvain Gubian \email{DL.RSupport@pmi.com}
}

