% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmltemplate.R
\name{getHTMLContainerCode}
\alias{getHTMLContainerCode}
\title{Generate a HTML table node code for component based on template}
\usage{
getHTMLContainerCode(id, toolParam)
}
\arguments{
\item{id}{String for component identification}

\item{toolParam}{\code{list} containing options for tools options
of the GraphRender component}
}
\value{
String which is the HTML code generated
}
\description{
Generate a HTML table node code for component based on template
}
\author{
Sylvain Gubian \email{DL.RSupport@pmi.com}
}

