% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htmltemplate.R
\name{getHTMLStyleCode}
\alias{getHTMLStyleCode}
\title{Generate a HTML style code for component based on template}
\usage{
getHTMLStyleCode(id)
}
\arguments{
\item{id}{String for component identification}
}
\value{
String which is the HTML style code generated
}
\description{
Generate a HTML style code for component based on template
}
\author{
Sylvain Gubian \email{DL.RSupport@pmi.com}
}

