% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jstemplate.R
\name{getJSCode}
\alias{getJSCode}
\title{Generate javascript code based on general options
and network data}
\usage{
getJSCode(dataJson, id, opts, toolParam)
}
\arguments{
\item{dataJson}{\code{list} containing network data for nodes and links}

\item{id}{String for component identification}

\item{opts}{\code{list} containing general options for GraphRender component}

\item{toolParam}{\code{list} containing urls to jquery, jquery-ui, d3js,
GraphRender JS library, options for the GraphRender tool}
}
\value{
String corresponding to JS code
}
\description{
Generate javascript code based on general options
and network data
}
\author{
Sylvain Gubian \email{DL.RSupport@pmi.com}
}

