% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getUUID}
\alias{getUUID}
\title{Function wich generates a UUID version 4}
\usage{
getUUID(seed = NULL)
}
\arguments{
\item{seed}{Integer for seeding the R random generator}
}
\value{
String corresponding to the UUID generated.
}
\description{
Function wich generates a UUID version 4
}

